refactor task batch env update onto JobDefn class
