+++ cirrus/cli/resources/__init__.py
@@ -1,3 +1,3 @@
 from .base import BaseResource
-from .resources import Resource, TaskResource
+from .resources import Resource
 from .outputs import Output
